+++ payments-core/src/main/java/com/stripe/android/model/PaymentMethodCreateParams.kt
@@ -1490,6 +1490,21 @@ constructor(
 
         @JvmStatic
         @JvmOverloads
+        fun createNgUssd(
+            billingDetails: PaymentMethod.BillingDetails? = null,
+            metadata: Map<String, String>? = null,
+            allowRedisplay: PaymentMethod.AllowRedisplay? = null,
+        ): PaymentMethodCreateParams {
+            return PaymentMethodCreateParams(
+                type = PaymentMethod.Type.NgUssd,
+                billingDetails = billingDetails,
+                metadata = metadata,
+                allowRedisplay = allowRedisplay,
+            )
+        }
+
+        @JvmStatic
+        @JvmOverloads
         fun createNgBankTransfer(
             billingDetails: PaymentMethod.BillingDetails? = null,
             metadata: Map<String, String>? = null,